+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+## [3.0.0]
+
+### Breaking Changes
+
+* use composition instead of inheritance with `S3Client` to behave like a plugin
+* move `@aws-sdk/client-s3` and `@aws-sdk/abort-controller` from dependencies to peer dependencies
+* deprecated option `flatten` has been removed, use `relocations` option instead
+
+### Features
+
+* add aws cli ``--include`` and ``--exclude`` support through `filters` option
+* replace promise chunks with promise queues for increased concurrency and performances
+* support multipart uploads when synchronizing a remote bucket with a local file system
+
+### Test
+
+* replace `mocha` with `jest`
+
 ## [2.1.0]
 
 * support aws cli ``--size-only`` option through `sizeOnly` property